Initial Commit
Its really early in the morning and I got it working :)
Added SiteInformation object to store data.
Added a site scraper to scrape a site on id passed in and create a SiteInformation object.
Added FileUtils  to create a csv file for the list on sites that is created.
Added CoroutineExtensions for pmap for parallel execution.